Extract tracking logic into a separate class
